requirements.txt
kivy
kivymd
tensorflow
numpy
pillow
pyinstaller
